[1.14] Rename folders
