+++ src/main/java/fr/epicanard/globalmarketchest/commands/consumers/ListConsumer.java
@@ -10,7 +10,7 @@ import fr.epicanard.globalmarketchest.commands.CommandConsumer;
 import fr.epicanard.globalmarketchest.commands.CommandNode;
 import fr.epicanard.globalmarketchest.permissions.Permissions;
 import fr.epicanard.globalmarketchest.shops.ShopInfo;
-import fr.epicanard.globalmarketchest.utils.ChatUtils;
+import fr.epicanard.globalmarketchest.utils.chat.ChatUtils;
 import fr.epicanard.globalmarketchest.utils.LangUtils;
 import net.md_5.bungee.api.ChatColor;
 import net.md_5.bungee.api.chat.TextComponent;
